[Task #1652] Revert renaming package assignment to occurrence.
