+++ src/main/java/cz/cvut/kbss/termit/dto/assignment/AbstractAssignmentInfo.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package cz.cvut.kbss.termit.dto.occurrence;
+package cz.cvut.kbss.termit.dto.assignment;
 
 import cz.cvut.kbss.jopa.model.annotations.OWLObjectProperty;
 import cz.cvut.kbss.jopa.model.annotations.Types;
