+++ src/main/java/cz/cvut/kbss/termit/service/document/AnnotationGenerator.java
@@ -20,8 +20,8 @@ package cz.cvut.kbss.termit.service.document;
 import cz.cvut.kbss.termit.exception.AnnotationGenerationException;
 import cz.cvut.kbss.termit.model.Asset;
 import cz.cvut.kbss.termit.model.Term;
-import cz.cvut.kbss.termit.model.occurrence.OccurrenceTarget;
-import cz.cvut.kbss.termit.model.occurrence.TermOccurrence;
+import cz.cvut.kbss.termit.model.assignment.OccurrenceTarget;
+import cz.cvut.kbss.termit.model.assignment.TermOccurrence;
 import cz.cvut.kbss.termit.model.resource.File;
 import cz.cvut.kbss.termit.model.selector.Selector;
 import cz.cvut.kbss.termit.persistence.dao.TermOccurrenceDao;
