+++ src/main/java/cz/cvut/kbss/termit/service/document/TermOccurrenceResolver.java
@@ -19,7 +19,7 @@ package cz.cvut.kbss.termit.service.document;
 
 import cz.cvut.kbss.termit.model.Asset;
 import cz.cvut.kbss.termit.model.Term;
-import cz.cvut.kbss.termit.model.occurrence.*;
+import cz.cvut.kbss.termit.model.assignment.*;
 import cz.cvut.kbss.termit.model.resource.File;
 import cz.cvut.kbss.termit.service.repository.TermRepositoryService;
